fix(chat): refactor - move common/buttons to components/buttons dir
