+++ apps/chat/src/components/Header/Header.tsx
@@ -15,7 +15,7 @@ import {
   OVERLAY_HEADER_ICON_SIZE,
 } from '@/src/constants/default-ui-settings';
 
-import { ToggleSidebarButton } from '@/src/components/Common/Buttons/ToggleSidebarButton';
+import { ToggleSidebarButton } from '@/src/components/Buttons/ToggleSidebarButton';
 import { SettingDialog } from '@/src/components/Settings/SettingDialog';
 
 import { BaseHeader } from './BaseHeader';
